@@ -19,7 +19,7 @@ public class TodoController {
 
     @GetMapping("/todos")
     public List<Todo> getAllTodos() {
-        Sort sortByCreatedAtDesc = new Sort(Sort.Direction.DESC, "createdAt");
+        Sort sortByCreatedAtDesc = Sort.by(Sort.Direction.DESC, "createdAt");
         return todoRepository.findAll(sortByCreatedAtDesc);
     }
 
